Revert "aout: shrink table storage"
This reverts commit 366cb7766235f9d5b7221dca033079a6c3c74036.

+++ include/vlc_aout.h
@@ -291,7 +291,7 @@ static_assert(AOUT_CHANIDX_MAX == AOUT_CHAN_MAX, "channel count mismatch");
 /**
  * It describes the audio channel order VLC expect.
  */
-static const uint16_t pi_vlc_chan_order_wg4[] =
+static const uint32_t pi_vlc_chan_order_wg4[] =
 {
     AOUT_CHAN_LEFT, AOUT_CHAN_RIGHT,
     AOUT_CHAN_MIDDLELEFT, AOUT_CHAN_MIDDLERIGHT,
